requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
faiss-cpu
numpy
pypdf
streamlit
requests
pydantic
python-dateutil
rank_bm25